Modify 'test' folder name
