+++ src/commands/__test__/tarea.test.ts
